requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
pytest==7.2.0
